Reworked Autoassigner & Hero's Will & Trade + Visual NX + New commands
Reworked autoassigner (improved limits between required secondary and
surplus primary stats). Hero's will removes most of diseases, tonic
removes slow. Added visual info for collected NX cards. Added commands
cake (cake boss with customizable HP) and setgmlevel. Reworked Trade
system now checking for slots smartly (instead of just checking for
empty slots).

+++ src/client/inventory/ItemFactory.java
@@ -80,7 +80,7 @@ public enum ItemFactory {
                 MapleInventoryType mit = MapleInventoryType.getByType(rs.getByte("inventorytype"));
 
                 if (mit.equals(MapleInventoryType.EQUIP) || mit.equals(MapleInventoryType.EQUIPPED)) {
-                    Equip equip = new Equip(rs.getInt("itemid"), (byte) rs.getInt("position"));
+                    Equip equip = new Equip(rs.getInt("itemid"), (short) rs.getInt("position"));
                     equip.setOwner(rs.getString("owner"));
                     equip.setQuantity((short) rs.getInt("quantity"));
                     equip.setAcc((short) rs.getInt("acc"));
